api/pyproject.toml
[tool.poetry]
name = "brandos-ai-api"
version = "0.1.0"
description = "BrandOS.AI API"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
